bytemap failing one amdgcn test, start considering how to model lack of pcie

+++ hostrpc/codegen/client.cpp
@@ -10,7 +10,7 @@ struct copy_functor_nop
 };
 
 using client_type =
-    hostrpc::client_impl<uint32_t, SZ, copy_functor_nop, hostrpc::nop_stepper>;
+    hostrpc::client_impl<uint64_t, SZ, copy_functor_nop, hostrpc::nop_stepper>;
 
 extern "C" __attribute__((noinline)) void client_instance_direct(client_type& c)
 {
